third_party/blink/web_tests/fast/events/window-onerror-08-expected.txt

PASS window.event.type is "error"
PASS onerrorMessage is window.event.message
PASS stripURL(onerrorURL) is stripURL(window.event.filename)
PASS onerrorLine is window.event.lineno
PASS onerrorColumn is window.event.colno
PASS successfullyParsed is true

TEST COMPLETE


This tests that when 'window.onerror' handler is called, 'window.event' is the corresponding ErrorEvent object.
